Add missing explicit cast from size_t to int to ssl_write_real()

+++ library/ssl_msg.c
@@ -5486,7 +5486,7 @@ static int ssl_write_real( mbedtls_ssl_context *ssl,
         return( ret );
     }
 
-    return( len );
+    return( (int) len );
 }
 
 /*
